refactor(backend): relocate plugin installer in separate csproj with tests
