+++ ScriptBeeWebApp/src/Plugins/Plugins.Installer/BundlePluginInstaller.cs
