+++ ScriptBeeWebApp/src/Plugins/Plugins.Installer/DownloadService.cs
