+++ ScriptBeeWebApp/src/Plugins/Plugins.Installer/IBundlePluginInstaller.cs
