+++ ScriptBeeWebApp/src/Plugins/Plugins.Installer/IBundlePluginUninstaller.cs
