+++ ScriptBeeWebApp/src/Plugins/Plugins.Installer/IDownloadService.cs
