+++ ScriptBeeWebApp/src/Plugins/Plugins.Installer/IPluginUninstaller.cs
