+++ ScriptBeeWebApp/src/Plugins/Plugins.Installer/ISimplePluginInstaller.cs
