+++ ScriptBeeWebApp/src/Plugins/Plugins.Installer/IZipFileService.cs
