+++ ScriptBeeWebApp/src/Plugins/Plugins.Installer/PluginInstallationError.cs
