+++ ScriptBeeWebApp/src/Plugins/Plugins.Installer/PluginVersionExistsError.cs
