+++ ScriptBeeWebApp/src/Plugins/Plugins.Installer/ZipFileService.cs
